+++ packages/next/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/next
 
+## 4.2.1
+
+### Patch Changes
+
+- [next] add streaming prerender group for actions ([#11454](https://github.com/vercel/vercel/pull/11454))
+
+- [next] Ensure \_not-found is included properly in app router functions ([#11441](https://github.com/vercel/vercel/pull/11441))
+
+- [next] add .action handling for dynamic routes ([#11461](https://github.com/vercel/vercel/pull/11461))
+
+- [next] Remove un-necessary compressed function size calc/constraint ([#11442](https://github.com/vercel/vercel/pull/11442))
+
 ## 4.2.0
 
 ### Minor Changes